CardsClassifier/requirements.txt
flask
torch
torchvision
efficientnet_pytorch
pillow
numpy
timm